:sparkles: Session api

+++ src/data/session/session.js
@@ -2,16 +2,77 @@ import { uriSession } from "./server";
 import axios from "axios";
 
 
+const FORBIDDEN=403;
+const BAD_REQUEST=400;
+const CONFLICT=409;
+const INTERNAL_ERROR = 500;
+const errorManager=(statusCode)=>{
+    switch (statusCode) {
+        case FORBIDDEN:
+            return FORBIDDEN + " No autorizado para hacer ejecutar esta petición";
+            break;
+        case BAD_REQUEST:
+            return BAD_REQUEST + " Falta algun campo o los campos no son validos";
+            break;
+        case CONFLICT:
+            return CONFLICT + " Se entró en conflico con la base de datos";
+            break;
+        case INTERNAL_ERROR:
+            return INTERNAL_ERROR + " Error interno";
+            break;
+        default:
+            return CONFLICT + " Error no identificado";
+    }
+
+}
+
+const IsDateright=(date)=>{
+    let correct = /^(20[0-9][0-9])-(0[0-9]|1[0-2])-([0-2][0-9]|3[0-1])?$/.test(date); 
+    if(correct){
+        const month = date.substring(5,7)
+        if(/^(04|06|09|11)?$/.test(month)){
+            correct = /^(20[0-9][0-9])-(0[0-9]|1[0-2])-([0-2][0-9]|30)?$/.test(date); 
+        }
+        if (month==="02"){
+            correct = /^(20[0-9][0-9])-(0[0-9]|1[0-2])-([0-2][0-8])?$/.test(date); 
+        }
+    }
+    return correct;
+}
+
+
+const IsTimeright=(time)=>{
+    const correct = /^([0-1][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9])?$/.test(time); 
+    return correct;
+}
+
+
 /*
 CREATE
 */
 
-export const registerSchedule = async (schedule) => {
+export const registerSchedule = async (schedule, Type) => {
     try {
+
+        if(Type===2){
+            const err = errorManager(FORBIDDEN);
+            throw new Error(err);
+        }
+        if(!IsDateright(schedule.daySession)|| 
+            !IsTimeright(schedule.iniTime)|| !IsTimeright(schedule.endTime)){
+                
+            const err = errorManager(BAD_REQUEST);
+            throw new Error(err);
+        }
         const response = await axios.post(uriSession + '/schedule/create',schedule);
-        return response.data;
+        return response.status;
     } catch (error) {
-        console.error(error);
+        if(error.response!=null){
+            const err = errorManager(error.response.status);
+            throw new Error(err);
+        } else{
+            throw error;
+        }
     }
 }
 
@@ -21,43 +82,65 @@ export const registerSchedule = async (schedule) => {
 DELETE
 */
 
-export const deleteSchedule = async (ChageState) => {
+export const deleteSchedule = async (ChageState, Type) => {
     try {
+        
+        if(Type===2){
+            const err = errorManager(FORBIDDEN);
+            throw new Error(err);
+        }
+
         const response = await axios.delete(uriSession + '/schedule/delete',{ data: ChageState});
-        return response.data;
+        return response.status;
     } catch (error) {
-        console.error(error);
+        if(error.response!=null){
+            const err = errorManager(error.response.status);
+            throw new Error(err);
+        } else{
+            throw error;
+        }
     }
 }
 
 /*
 PUT
 */
-export const setAdate = async (ChageState) => {
+export const setAdate = async (ChageState, Type) => {
     try {
+        if(Type===1){
+            const err = errorManager(FORBIDDEN);
+            throw new Error(err);
+        }
+
         const response = await axios.put(uriSession + '/set-a-date',ChageState);
-        return response.data;
+        return response.status;
     } catch (error) {
-        console.error(error);
+        if(error.response!=null){
+            const err = errorManager(error.response.status);
+            throw new Error(err);
+        } else{
+            throw error;
+        }
     }
 }
 
 export const calcelUser = async (ChageState) => {
     try {
         const response = await axios.put(uriSession + '/cancel/user',ChageState);
-        return response.data;
+        return response.status;
     } catch (error) {
-        console.error(error);
-        throw new error( error.response.status+" "+ error.response.statusCode );
+        const err = errorManager(error.response.status)
+        throw new Error(err);
     }
 }
 
 export const calcelCoach = async (ChageState) => {
     try {
         const response = await axios.put(uriSession + '/cancel/coach',ChageState);
-        return response.data;
+        return response.status;
     } catch (error) {
-        console.error(error);
+        const err = errorManager(error.response.status)
+        throw new Error(err);
     }
 }
 
@@ -69,7 +152,8 @@ export const getAllbyIdCoach = async (idCoach) => {
         const response = await axios.get(uriSession + '/get-by-idCoach/' + idCoach);
         return response.data;
     } catch (error) {
-        console.error(error);
+        const err = errorManager(error.response.status);
+        throw new Error(err);
     }
 }
 export const getAllbyCoachCurrent = async (idCoach) => {
@@ -77,7 +161,17 @@ export const getAllbyCoachCurrent = async (idCoach) => {
         const response = await axios.get(uriSession + '/get-by-idCoach/Current/' + idCoach);
         return response.data;
     } catch (error) {
-        console.error(error);
+        const err = errorManager(error.response.status);
+        throw new Error(err);
+    }
+}
+export const getAllbyCoachAvaible = async (idCoach) => {
+    try {
+        const response = await axios.get(uriSession + '/get-by-idCoach/Avaible/' + idCoach);
+        return response.data;
+    } catch (error) {
+        const err = errorManager(error.response.status);
+        throw new Error(err);
     }
 }
 
@@ -86,7 +180,8 @@ export const getAllbyIdUser = async (idUser) => {
         const response = await axios.get(uriSession + '/get-by-idUser/' + idUser);
         return response.data;
     } catch (error) {
-        console.error(error);
+        const err = errorManager(error.response.status);
+        throw new Error(err);
     }
 }
 export const getAllbyUserCurrent = async (idUser) => {
@@ -94,6 +189,18 @@ export const getAllbyUserCurrent = async (idUser) => {
         const response = await axios.get(uriSession + '/get-by-idUser/Current/' + idUser);
         return response.data;
     } catch (error) {
-        console.error(error);
+        const err = errorManager(error.response.status);
+        throw new Error(err);
     }
 }
+export const getbyIdSchedule = async (idUser) => {
+    try {
+        const response = await axios.get(uriSession + '/get-by-idSchedule/' + idUser);
+        return response.data;
+    } catch (error) {
+        const err = errorManager(error.response.status);
+        throw new Error(err);
+        
+    }
+}
+
